feat: 🎸 added system notification
✅ Closes: #393
