+++ src-tauri/tauri.conf.json
@@ -49,6 +49,9 @@
         },
         "allowlist": {
             "all": true,
+            "notification": {
+                "all": true
+            },
             "http": {
                 "scope": [
                     "https://manjaro.org/search/*",
