+++ src/translations/en.json
@@ -133,8 +133,8 @@
         "openPamacTooltip": "It opens pamac update page.",
         "updateTooltip": "It runs pamac update command.",
         "commandHistory": "Command History",
-        "installSuccess": "Installed successfully.",
-        "installError": "Failed installed.",
+        "installSuccess": "installed successfully.",
+        "installError": "installation failed.",
         "failed": "Failed",
         "success": "Success",
         "learnMore": "Learn More",
