+++ src/utils/StringHelper.ts
@@ -1,0 +1,4 @@
+function capitalize(str:string) {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
+export default capitalize;
